refactor: extract CopyableBox and MasterKeyDialog shared components
Move CopyableBox (with hasActiveSelection guard) from AgentApiKeysSection
into a standalone shared component so the guard is always present for any
caller. Extract MasterKeyDialog using CopyableBox, replacing the inline
dialog in Sidebar and NetworksPage with a single shared component.
