+++ frontend/src/app/networks/page.tsx
@@ -1,13 +1,13 @@
 import { useCallback, useState } from 'react';
 import { useNavigate } from 'react-router';
 import * as Tabs from '@radix-ui/react-tabs';
-import * as Dialog from '@radix-ui/react-dialog';
-import { Plus, Users, Loader2, X, Copy, Check } from 'lucide-react';
+import { Plus, Users, Loader2 } from 'lucide-react';
 import NetworkAvatar from '@/components/IndexAvatar';
 import ClientLayout from '@/components/ClientLayout';
 import CreateNetworkModal from '@/components/modals/CreateIndexModal';
 import { ContentContainer } from '@/components/layout';
 import { Button } from '@/components/ui/button';
+import MasterKeyDialog from '@/components/MasterKeyDialog';
 import { useAuthContext } from '@/contexts/AuthContext';
 import { useNotifications } from '@/contexts/NotificationContext';
 import { useNetworks } from '@/contexts/APIContext';
@@ -27,7 +27,6 @@ export default function NetworksPage() {
   const [loadingPublic, setLoadingPublic] = useState(false);
   const [joiningNetwork, setJoiningNetwork] = useState<string | null>(null);
   const [masterKeyModal, setMasterKeyModal] = useState<{ networkId: string; masterKey: string } | null>(null);
-  const [copied, setCopied] = useState(false);
 
   const allNetworks = [...(rawIndexes || [])].filter(Boolean).sort((a, b) => {
     if (a.isPersonal && !b.isPersonal) return -1;
@@ -232,69 +231,15 @@ export default function NetworksPage() {
         uploadIndexImage={indexesService.uploadIndexImage}
       />
 
-      <Dialog.Root open={!!masterKeyModal} onOpenChange={(open) => {
-        if (!open) {
+      <MasterKeyDialog
+        open={!!masterKeyModal}
+        masterKey={masterKeyModal?.masterKey ?? ''}
+        onClose={() => {
           const networkId = masterKeyModal?.networkId;
           setMasterKeyModal(null);
-          setCopied(false);
           if (networkId) navigate(`/networks/${networkId}`);
-        }
-      }}>
-        <Dialog.Portal>
-          <Dialog.Overlay className="fixed inset-0 bg-black/50 z-[100]" />
-          <Dialog.Content className="fixed left-1/2 top-1/2 -translate-x-1/2 -translate-y-1/2 bg-white rounded-sm shadow-lg w-full max-w-md z-[100] focus:outline-none">
-            <div className="p-6">
-              <div className="flex items-center justify-between mb-4">
-                <Dialog.Title className="text-lg font-bold text-black">Master Key</Dialog.Title>
-                <Dialog.Close className="p-1 rounded-sm hover:bg-gray-100 text-gray-400 hover:text-gray-600 transition-colors">
-                  <X className="h-4 w-4" />
-                </Dialog.Close>
-              </div>
-              <p className="text-sm text-gray-600 mb-4">
-                Save this key now — it will not be shown again. Use it as the <code className="text-xs bg-gray-100 px-1 py-0.5 rounded">x-api-key</code> header when calling the signup endpoint.
-              </p>
-              <button
-                type="button"
-                onClick={async () => {
-                  if (!masterKeyModal?.masterKey) return;
-                  const sel = window.getSelection();
-                  if (sel && !sel.isCollapsed) return;
-                  try {
-                    await navigator.clipboard.writeText(masterKeyModal.masterKey);
-                    setCopied(true);
-                    setTimeout(() => setCopied(false), 800);
-                  } catch { /* silent */ }
-                }}
-                aria-label="Copy"
-                className={`relative w-full text-left group rounded-sm border p-3 transition-colors duration-300 ${
-                  copied
-                    ? 'bg-green-100 border-green-400'
-                    : 'bg-gray-50 border-gray-200 hover:bg-green-50 hover:border-green-300'
-                }`}
-              >
-                <code className="block text-xs text-gray-700 font-ibm-plex-mono whitespace-pre-wrap break-all pr-16 select-text">{masterKeyModal?.masterKey}</code>
-                <span className="absolute top-2 right-2 inline-flex items-center gap-1 text-xs text-gray-400 group-hover:text-green-700 transition-colors select-none">
-                  {copied ? (
-                    <><Check className="w-3 h-3" /> Copied</>
-                  ) : (
-                    <><Copy className="w-3 h-3" /> Copy</>
-                  )}
-                </span>
-              </button>
-              <div className="flex justify-end mt-4">
-                <Button onClick={() => {
-                  const networkId = masterKeyModal?.networkId;
-                  setMasterKeyModal(null);
-                  setCopied(false);
-                  if (networkId) navigate(`/networks/${networkId}`);
-                }}>
-                  Done
-                </Button>
-              </div>
-            </div>
-          </Dialog.Content>
-        </Dialog.Portal>
-      </Dialog.Root>
+        }}
+      />
     </ClientLayout>
   );
 }
